requirements.txt
django-crispy-forms==1.7.2
django-markdownify==0.8.0
Django==2.1.5
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python-decouple==3.1
celery==4.2.1
celery[redis]
django-celery-results==1.0.4
djangorestframework==3.9.1
